refactor : 패키지 구조 개선 및 enum 추가
