+++ src/main/kotlin/ac/dnd/server/enums/ApplicantType.kt
@@ -2,6 +2,8 @@ package ac.dnd.server.enums
 
 enum class ApplicantType(val description: String) {
     FRONTEND("프론트엔드"),
+    ANDROID("안드로이드"),
+    IOS("iOS"),
     BACKEND("백엔드"),
     DESIGNER("디자이너")
 }
