+++ src/main/kotlin/ac/dnd/server/global/config/properties/PropertiesSourceConfig.kt
@@ -1,8 +1,8 @@
-package ac.dnd.server.config.properties
+package ac.dnd.server.global.config.properties
 
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-@ConfigurationPropertiesScan(basePackages = ["ac.dnd.server.config.properties"])
+@ConfigurationPropertiesScan(basePackages = ["ac.dnd.server.global.config.properties"])
 class PropertiesSourceConfig
